+++ previous-versions/assess/pkg/models/operations/generateloantransactions.go
@@ -64,7 +64,7 @@ func (o *GenerateLoanTransactionsRequest) GetSourceType() GenerateLoanTransactio
 type GenerateLoanTransactionsResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
+	// The request made is not valid.
 	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int