+++ previous-versions/assess/pkg/models/operations/getaccountingmarketingmetrics.go
@@ -85,7 +85,7 @@ func (o *GetAccountingMarketingMetricsRequest) GetShowInputValues() *bool {
 type GetAccountingMarketingMetricsResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
+	// The request made is not valid.
 	ErrorMessage *shared.ErrorMessage
 	// OK
 	Report *shared.Report