+++ previous-versions/assess/pkg/models/operations/getcommerceordersmetrics.go
@@ -76,7 +76,7 @@ func (o *GetCommerceOrdersMetricsRequest) GetReportDate() string {
 type GetCommerceOrdersMetricsResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
+	// The request made is not valid.
 	ErrorMessage *shared.ErrorMessage
 	// OK
 	Report *shared.Report