+++ previous-versions/assess/pkg/models/operations/getdataintegritysummaries.go
@@ -40,7 +40,7 @@ func (o *GetDataIntegritySummariesRequest) GetQuery() *string {
 type GetDataIntegritySummariesResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
+	// Your `query` parameter was not correctly formed
 	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int