+++ previous-versions/assess/pkg/models/operations/listdatatypedataintegritydetails.go
@@ -81,7 +81,7 @@ type ListDataTypeDataIntegrityDetailsResponse struct {
 	ContentType string
 	// OK
 	Details *shared.Details
-	// Your API request was not properly authorized.
+	// Your `query` parameter was not correctly formed
 	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int